AACT-142:  Simplify & clean up code a little bit.
